requirements.txt
pandas==0.25.3
python-telegram-bot==12.2.0
matplotlib==3.1.2
geopy==1.20.0
requests==2.22.0
